decision making

+++ Section 2 Decision Making (21–40)/21. Even or Odd.py
@@ -4,7 +4,7 @@
 # Author: Anugya Agrawal
 # ---------------------------------------
 
-number = float(input("Enter Number: "))
+number = int(input("Enter Number: "))
 
 if number % 2 == 0:
     print(number,"Number is Even")
